+++ src/ServerAPI.cpp
@@ -104,7 +104,7 @@ std::string ServerAPI::getSecondaryUrl() {
 void ServerAPI::init() {
     m_amazon = evaluateAmazon();
     static_assert(GEODE_COMP_GD_VERSION == REQUIRED_GD_VERSION, "Unsupported GD version");
-    static_assert(isSupportedPlatform(), "Unsupported platform");
+    static_assert(IS_SUPPORTED_PLATFORM, "Unsupported platform");
 
     this->m_baseUrl = (char*)(geode::base::get() + (m_amazon ? BASE_URL_OFFSET_AMAZON : BASE_URL_OFFSET));
     this->m_secondaryUrl = ZipUtils::base64URLDecode((char *)(geode::base::get() + (m_amazon ? SECONDARY_URL_OFFSET_AMAZON : SECONDARY_URL_OFFSET)));
@@ -156,7 +156,8 @@ decltype(auto) ServerAPI::doAndNotifyIfServerUpdate(CFunc&& func, Mod* updater, 
 
     // Server has updated
     if (oldID != m_cache.ID) {
-        ServerUpdatingEvent(updater, m_cache.URL).post();
+        ServerUpdatingEventData event(updater, m_cache.URL);
+        ServerUpdatingEvent().send(&event);
     }
 
     // Cast back to void if original result type is void
@@ -168,26 +169,26 @@ bool ServerAPI::isAmazon() {
 }
 
 $on_mod(Loaded) {
-    new EventListener<EventFilter<GetCurrentServerEvent>>(+[](GetCurrentServerEvent* e) {
+    GetCurrentServerEvent().listen([](GetCurrentServerEventData* e) {
         e->m_url = ServerAPI::get()->getCurrentURL();
         e->m_prio = ServerAPI::get()->getCurrentPrio();
         e->m_id = ServerAPI::get()->getCurrentId();
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<GetServerByIdEvent>>(+[](GetServerByIdEvent* e) {
+    GetServerByIdEvent().listen([](GetServerByIdEventData* e) {
         e->m_url = ServerAPI::get()->getURLById(e->m_id);
         e->m_prio = ServerAPI::get()->getPrioById(e->m_id);
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<RegisterServerEvent>>(+[](RegisterServerEvent* e) {
+    RegisterServerEvent().listen([](RegisterServerEventData* e) {
         //e->m_id = ServerAPI::get()->registerURL(e->m_url, e->m_priority);
         e->m_id = ServerAPI::get()->doAndNotifyIfServerUpdate(&ServerAPI::registerURL, e->sender, e->m_url, e->m_priority);
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<UpdateServerEvent>>(+[](UpdateServerEvent* e) {
+    UpdateServerEvent().listen([](UpdateServerEventData* e) {
         if (e->m_url != "") {
             if (e->m_priority != 0) {
                 //ServerAPI::get()->updateURLAndPrio(e->m_id, e->m_url, e->m_priority);
@@ -201,31 +202,31 @@ $on_mod(Loaded) {
             ServerAPI::get()->doAndNotifyIfServerUpdate(&ServerAPI::updatePrio, e->sender, e->m_id, e->m_priority);
         }
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<RemoveServerEvent>>(+[](RemoveServerEvent* e) {
+    RemoveServerEvent().listen([](RemoveServerEventData* e) {
         //ServerAPI::get()->removeURL(e->m_id);
         ServerAPI::get()->doAndNotifyIfServerUpdate(&ServerAPI::removeURL, e->sender, e->m_id);
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<GetBaseUrlEvent>>(+[](GetBaseUrlEvent* e) {
+    GetBaseUrlEvent().listen([](GetBaseUrlEventData* e) {
         e->m_url = ServerAPI::get()->getBaseUrl();
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<GetSecondaryUrlEvent>>(+[](GetSecondaryUrlEvent* e) {
+    GetSecondaryUrlEvent().listen([](GetSecondaryUrlEventData* e) {
         e->m_url = ServerAPI::get()->getSecondaryUrl();
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<GetRegisteredServersEvent>>(+[](GetRegisteredServersEvent* e) {
+    GetRegisteredServersEvent().listen([](GetRegisteredServersEventData* e) {
         e->m_servers = ServerAPI::get()->getAllServers();
         return ListenerResult::Stop;
-    });
+    }).leak();
 
-    new EventListener<EventFilter<GetTrustedModsEvent>>(+[](GetTrustedModsEvent* e) {
+    GetTrustedModsEvent().listen([](GetTrustedModsEventData* e) {
         e->trustedMods = ServerAPI::get()->getTrustedMods();
         return ListenerResult::Stop;
-    });
+    }).leak();
 }